[update] インベントリ操作の代数にcreateを追加する
